+++ test/false/expected.js
@@ -1,10 +1,10 @@
-'use strict';
+"use strict";
 
 var x = false;
 
 if (!false) {
-  console.log('Debug info');
+  console.log("Debug info");
 }
 if (false) {
-  console.log('Production log');
+  console.log("Production log");
 }