+++ test/member-expression/actual.js
@@ -1,2 +1,4 @@
+"use strict";
+
 process.env.NODE_ENV;
 process.env.NODE_ENV === "development";