+++ test/null/actual.js
@@ -1,8 +1,11 @@
+"use strict";
+
 var x = PRODUCTION;
 
 if (!PRODUCTION) {
-  console.log('Debug info');
+  console.log("Debug info");
 }
+
 if (PRODUCTION) {
-  console.log('Production log');
+  console.log("Production log");
 }