+++ test/unary-expression/actual.js
@@ -1,2 +1,4 @@
+"use strict";
+
 typeof window;
 typeof window === "object";