+++ test/undefined/expected.js
@@ -1,10 +1,11 @@
-'use strict';
+"use strict";
 
 var x = undefined;
 
 if (!undefined) {
-  console.log('Debug info');
+  console.log("Debug info");
 }
+
 if (undefined) {
-  console.log('Production log');
+  console.log("Production log");
 }