+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Change plugin options to **only** be a real JS object. Removes string configuration path option as now this is all possible with dynamic `.babelrc.js` or `babel.config.js` files.
 * Update to `@babel/core` / Babel 7+.
 
-#### Infrastructure
+#### Internal
 
 * Lint all `test` code.